+++ src/io/tcp_server.cpp
@@ -156,7 +156,7 @@ namespace YanLib::io {
     }
 
     std::string tcp_server::read_string_to_end(SOCKET client_socket) {
-        constexpr DWORD buffer_size = 4096;
+        constexpr unsigned long buffer_size = 4096;
         char *buf = new char[buffer_size];
         memset(buf, 0, buffer_size);
         std::string raw_data;
