+++ src/io/udp_client.cpp
@@ -188,7 +188,7 @@ namespace YanLib::io {
 
     std::string udp_client::read_string_to_end(std::string &remote_ip,
                                                uint16_t &remote_port) {
-        constexpr DWORD buffer_size = 4096;
+        constexpr unsigned long buffer_size = 4096;
         char *buf = new char[buffer_size];
         memset(buf, 0, buffer_size);
         std::string raw_data;
